Avoid unwanted bean overrides in spring-boot-test-autoconfigure
See gh-13609
